#1535# Changed opt-in policy: Only readers are automatically added.

+++ classes/user/form/RegistrationForm.inc.php
@@ -210,9 +210,11 @@ class RegistrationForm extends Form {
 		// about journal updates. (The delete is here to prevent a
 		// duplicate insert error msg if there was a notification entry
 		// left over from a previous role.)
-		$notificationStatusDao = &DAORegistry::getDAO('NotificationStatusDAO');
-		$notificationStatusDao->deleteNotificationStatusByJournal($journal->getJournalId());
-		$notificationStatusDao->setJournalNotifications($journal->getJournalId(), $userId, true);
+		if (isset($allowedRoles['reader']) && $this->getData($allowedRoles['reader'])) {
+			$notificationStatusDao = &DAORegistry::getDAO('NotificationStatusDAO');
+			$notificationStatusDao->deleteNotificationStatusByJournal($journal->getJournalId());
+			$notificationStatusDao->setJournalNotifications($journal->getJournalId(), $userId, true);
+		}
 	}
 	
 }
